+++ src/utils/web3converter.py
@@ -2,6 +2,9 @@ from src.types import SlotNumber, EpochNumber, FrameNumber
 from src.modules.submodules.types import ChainConfig, FrameConfig
 
 
+def epoch_from_slot(slot: SlotNumber, slots_per_epoch: int) -> EpochNumber:
+    return EpochNumber(slot // slots_per_epoch)
+
 class Web3Converter:
     """
     The Web3Converter class contains methods for converting between slot, epoch, and frame numbers using chain and
